Simplify Project using ES6
