+++ src/component/base-components.ts
@@ -1,7 +1,5 @@
-namespace App {
-
-    // Component Base Class
- export abstract class Component<T extends HTMLElement, U extends HTMLElement> {
+// Component Base Class
+export abstract class Component<T extends HTMLElement, U extends HTMLElement> {
     templateElement: HTMLTemplateElement;
     hostElement: T ;
     element: U;
@@ -35,7 +33,5 @@ namespace App {
 
     abstract configure():void;
     abstract renderContent(): void;
-
-  }
 
 }
