+++ src/component/project-input.ts
@@ -1,11 +1,11 @@
-/// <reference path="base-components.ts" />
-/// <reference path="../decorators/autobind-decorator.ts"/>
-/// <reference path="../util/validation.ts"/>
-/// <reference path="../state/project-state.ts"/>
+import { Component } from './base-components.js';
+import { Validatable, validate } from '../util/validation.js'
+import { autobind } from '../decorators/autobind.js';
+import { projectState } from '../state/project-state.js';
 
-namespace App {
 
-    // ProjectInput Class
+
+// ProjectInput Class
     export class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
         titleInputElement: HTMLInputElement;
         descriptionInputElement: HTMLInputElement;
@@ -81,4 +81,3 @@ namespace App {
             }
         }
     }
-}
