+++ src/models/drag-drop.ts
@@ -1,17 +1,16 @@
 // Drag and Drop Interfaces
-namespace App {
 
-    export interface Draggable{
-        dragStartHandler(event: DragEvent): void
-        dragEndHandler(event: DragEvent): void
-    }
-
-    export interface DragTarget {
-        dragOverHandler(event:DragEvent): void;
-        dropHandler(event:DragEvent): void;
-        dragLeaveHandler(event:DragEvent): void;
-    }
-
+export interface Draggable{
+    dragStartHandler(event: DragEvent): void
+    dragEndHandler(event: DragEvent): void
 }
+
+export interface DragTarget {
+    dragOverHandler(event:DragEvent): void;
+    dropHandler(event:DragEvent): void;
+    dragLeaveHandler(event:DragEvent): void;
+}
+
+
