+++ src/models/project.ts
@@ -1,16 +1,14 @@
 // Project Type
+export enum ProjectStatus {active, finished}
 
-namespace App {
-    export enum ProjectStatus {active, finished}
-
-    export class Project {
-        constructor(
-        public id: string,
-        public title: string,
-        public description: string,
-        public people: number,
-        public status: ProjectStatus){
-        }
+export class Project {
+    constructor(
+    public id: string,
+    public title: string,
+    public description: string,
+    public people: number,
+    public status: ProjectStatus){
     }
-
 }
+
+
