+++ src/state/project-state.ts
@@ -1,4 +1,5 @@
-namespace App {
+import { Project } from "../models/project.js";
+import { ProjectStatus } from "../models/project.js";
 
 
 type Listener<T> = (items: T[]) => void
@@ -57,4 +58,3 @@ export class ProjectState extends State<Project> {
 
 export const projectState =  ProjectState.getInstance()
 
-}
