+++ src/util/validation.ts
@@ -1,6 +1,4 @@
 
-namespace App {
-
 // Validation
 export interface Validatable {
     value: string | number;
@@ -44,4 +42,3 @@ export interface Validatable {
     }
     return isValid;
   }
-}
